Upgrade jquery-migrate to 3.5.2
