+++ core/org.wso2.carbon.ui/src/main/resources/web/admin/layout/template.jsp
@@ -122,7 +122,7 @@
         }
     %>
     <script type="text/javascript" src="../admin/js/jquery-3.6.0.min.js"></script>
-    <script type="text/javascript" src="../admin/js/jquery-migrate-3.3.2.js"></script>
+    <script type="text/javascript" src="../admin/js/jquery-migrate-3.5.2.js"></script>
     <script type="text/javascript" src="../admin/js/jquery.form.js"></script>
     <script type="text/javascript" src="../dialog/js/jqueryui/jquery-ui.min.js"></script>
     <script type="text/javascript" src="../admin/js/jquery.validate.js"></script>
